check lab 10 prob 1 2

+++ src/lab10/problem1/prob1B/prob_ii.java
@@ -8,7 +8,7 @@ public class prob_ii {
         class RandomSupplier implements Supplier<Double> {
             @Override
             public Double get() {
-                return (Double) Math.random();
+                return Math.random();
             }
         }
